% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_content}
\alias{extract_content}
\title{Extracts the xml data}
\usage{
extract_content(yr_content)
}
\arguments{
\item{yr_content}{extracted data from yr.no}
}
\value{
a data frame
}
\description{
It is a non-exported helper function for get_weather().
}
