% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_creditation}
\alias{extract_creditation}
\title{Extracts the creditation}
\usage{
extract_creditation(yr_content)
}
\arguments{
\item{yr_content}{extracted data from yr.no}
}
\value{
a string
}
\description{
It is a non-exported helper function for get_weather().
}
