% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Function to retrieve and parse the xml forecast data from YR.no}
\usage{
get_weather(place_url, forecast = "days")
}
\arguments{
\item{place_url}{the url pointing to the place. Use search_places() to find
the right url.}

\item{forecast}{set to either "days" or "hours". Days returns six hour intervals
for several days. Hours returns hourly forecasts for the next 48 hours.}
}
\description{
Function to retrieve and parse the xml forecast data from YR.no
}
\examples{
\dontrun{
library(norweather)

places <- search_places("Copenhagen")
place_url <- places$place_url[1]

weather_data <- get_weather(place_url)

}
}
