% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prettify_data}
\alias{prettify_data}
\title{Makes the data frame pretty}
\usage{
prettify_data(extracted_content)
}
\arguments{
\item{extracted_content}{}
}
\value{
a data frame
}
\description{
It is a non-exported helper function for get_weather().
}
