% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_places.R
\name{search_places}
\alias{search_places}
\title{Search yr.no for a place.}
\usage{
search_places(place = "", country = "")
}
\arguments{
\item{place}{The place you want to search for. Keep the search as simple as
possible. This function will only return the first page of the search
result from yr.no}

\item{country}{Two letter country code. Ex: DK, US, DE.}
}
\description{
Search yr.no for a place.
}
\examples{

library(norweather)

search_places(place = "Copenhagen")
search_places(place = "New York")
search_places(place = "London")
}
