% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr.R
\name{yr}
\alias{yr}
\title{Function to retrieve and parse the xml forecast data from YR.no}
\usage{

  yr(xml_url = "http://www.yr.no/place/Denmark/Central_Jutland/Aarhus/forecast.xml",
  return_location = FALSE)
}
\arguments{
\item{xml_url}{the url pointing to the xml file.}

\item{return_location}{Boolean. If false it returns only the tidy data.frame of forecast data.}
}
\description{
Function to retrieve and parse the xml forecast data from YR.no
}
\examples{
library(dplyr)
library(yrno)

aarhus <- "http://www.yr.no/place/Denmark/Central_Jutland/Aarhus/forecast.xml"
cph <- "http://www.yr.no/sted/Danmark/Hovedstaden/København/varsel_time_for_time.xml"

yr(aarhus)
yr(cph)

}

