% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_bike_df.R
\name{yr_bike_df}
\alias{yr_bike_df}
\title{Get the data on the on the hours you want to ride your bike.}
\usage{
yr_bike_df(place = NULL, country = NULL, place_url = NULL, hour,
  forecast_date = Sys.Date(), hours_safety = 0)
}
\arguments{
\item{place}{City you want to ride your bike from.}

\item{place_url}{The place url. Ex http://www.yr.no/place/Denmark/Capital/Copenhagen}

\item{hour}{The hour you want to ride your bike.}

\item{forecast_date}{Default is current date.}

\item{hours_safety}{The safety margin subtracted and
to the hour param.}
}
\description{
If you want to ride your bike, but care about getting your feet
wet, then you need weather forecasts about the hours you expect
to ride your bike.
}
\examples{

yr_bike(place = "Copenhagen",
        country = "DK",
        hour = 7,
        forecast_data = Sys.Date() + 1,
        hours_safety = 1)
yr_bike(hour = 7, forecast_date = Sys.Date() + 1,
        hours_safety = 1,
        place_url = "https://www.yr.no/place/Denmark/Capital/Copenhagen/")
yr_bike(place_url = "https://www.yr.no/place/Denmark/Capital/Copenhagen/",
        hour = 7, forecast_date = Sys.Date() + 1,
        hours_safety = 1)
yr_bike(place_url = "https://www.yr.no/place/Denmark/Capital/Copenhagen/",
        hour = 7,
        forecast_date = Sys.Date() + 1,
        hours_safety = 1)

}
