% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_hour_by_hour.R
\name{yr_hour_by_hour}
\alias{yr_hour_by_hour}
\title{Get hourly forecast from YR.no}
\usage{
yr_hour_by_hour(place = NULL, country = NULL, place_url = NULL,
  search_no = 1)
}
\arguments{
\item{place}{Place name. Ex 'Copenhagen', 'Frederiksberg' 'New York'.
The place name is used to find the place_url by using the yr_search function.
The search results are dependent on your location and language, so if
you want to be certain that you get the correct forecast, then supply
a place_url that you get from either the yr_search function or directly
from the YR.no website.}

\item{country}{The two digit country code. Ex 'US', 'DK', 'SE'}

\item{place_url}{The place_url can be copied directly from the YR.no
website and used instead of the place param.}
}
\description{
Function to get the hourly forecast from YR.no, by supplying
a city name or a place_url copied from YR.no.
}
\examples{

library(yrno)

yr_hour_by_hour(place = "Copenhagen")
yr_hour_by_hour(place = "Copenhagen", country = "DK")
yr_hour_by_hour(place_url = yr_search(place = "Copenhagen")$url[1])

}
