% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_search.R
\name{yr_search}
\alias{yr_search}
\title{Search yr.no for a place.}
\usage{
yr_search(place, country = NULL)
}
\arguments{
\item{place}{The place you wan't to search for. Keep
the search as simple as possible. This function will
only return the fist page of the search result from
yr.no}

\item{country}{Two letter country code. Ex: DK, US, DE.}
}
\description{
Search yr.no for a place.
}
\examples{

library(yrno)

yr_search("Copenhagen")
yr_search("Copenhagen", "US")
yr_search("New York")
yr_search("London")

}

