% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_bike_recommendation.R
\name{yr_bike_recommendation}
\alias{yr_bike_recommendation}
\title{Figure out if you are going to be wet, if you ride your bike to work.}
\usage{
yr_bike_recommendation(place = NULL, country = NULL, place_url = NULL,
  hour, forecast_date = Sys.Date() + 1, hours_safety = 0, min_list = NULL,
  max_list = NULL, return_boolean = FALSE)
}
\arguments{
\item{place}{City you want to ride your bike from. It is recommended to
use the place_url instead of place (and country) as these params needs
more requests and webscraping to get to the final results.}

\item{country}{To two digit country code. Ex: "DK", "SE", "US".}

\item{place_url}{The page you get, when you search for a place on YR.no.
Ex http://www.yr.no/place/Denmark/Capital/Copenhagen}

\item{hour}{The hour you want to ride your bike.}

\item{forecast_date}{Default is current date.}

\item{hours_safety}{The safety margin subtracted and
added to the hour param.}

\item{return_boolean}{Do you want the function to return a boolean?}
}
\description{
Function to get the forecasts for when you are going to ride your
bike to and from work.
}
\examples{

yr_bike_recommendation("Copenhagen", 7, Sys.Date() + 1, hours_safety = 1)
yr_bike_recommendation(hour = 7, forecast_date = Sys.Date() + 1,
                hours_safety = 1,
                place_url = "https://www.yr.no/place/Denmark/Capital/Copenhagen/")
yr_bike_recommendation(place_url = "https://www.yr.no/place/Denmark/Capital/Copenhagen/",
                hour = 7, forecast_date = Sys.Date() + 1,
                hours_safety = 1,
                max_list = list(temperature_value = 5))
yr_bike_recommendation(place_url = "https://www.yr.no/place/Denmark/Capital/Copenhagen/",
                hour = 7, forecast_date = Sys.Date() + 1,
                hours_safety = 1,
                min_list = list(temperature_value = 5),
                max_list = list(precipitation_value = 0.1))

}

